server/requirements.txt
fastapi>=0.95,<1.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4,<2.0
aiosqlite>=0.18.0
python-dotenv>=0.21.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
pydantic>=1.10,<2.0
requests>=2.31.0
python-multipart>=0.0.6
pillow>=9.5.0
pytesseract>=0.3.10
jsonschema>=4.18.0
typing-extensions>=4.5.0
httpx==0.25.2
tenacity==9.1.2
aioredis==2.0.1
authlib==1.6.5
starlette-session==0.4.3